api/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
asyncpg==0.29.0
bcrypt==4.1.3
PyJWT==2.8.0
pydantic==2.7.1